fix(dedup): CR-11/CR-14/CR-18/CR-20–23 float timeout, zero guard, error_message cleanup
CR-18: Widen timeout_per_candidate from Optional[int] to Optional[float] across
AnalysisConfig, analysis_orchestrator (all method signatures and _resolve_timeout
return type), coverage._process_parallel_batch/_get_test_coverage_for_files_batch,
and tools.py. Enables sub-second timeouts in tests without 4s real sleeping.

CR-11: Guard timeout_per_candidate <= 0 in _resolve_timeout — treat as None and
return the default (previously timeout=0 would instantly fail every running future,
a latent bug activated by the now-live timeout enforcement).

CR-14: Replace error_message: Optional[str] smuggling in _handle_enrichment_error
with timeout_seconds: Optional[float]. The on_error callback now passes the numeric
timeout directly; _handle_enrichment_error formats with :g to strip trailing zeros
(1.0 → "1s", 0.2 → "0.2s"). The log field timeout_seconds is now a number, not
a string.

CR-20/CR-21/CR-22/CR-23: Hoist import threading / import time to module level in
test_orchestrator_optimizations.py; extract class constants TIMEOUT_SECONDS,
HUNG_WORKER_SAFETY_NET_SECONDS, ELAPSED_BOUND_SECONDS, EXPECTED_TIMEOUT_ERROR in
TestPerCandidateTimeoutEnforcement and TestSequentialPathTimeoutEnforcement; reduce
TIMEOUT_SECONDS from 1 → 0.2 (float) in both classes, cutting ~4s from the unit
suite. TestAbandonedWorkerIsolation retains TIMEOUT_SECONDS=1 (int) for its
worker-join timing margin.

CR-28 (partial): test_all_hung_candidates_each_marked_timed_out now asserts
"error" in c rather than pinning the exact label for never-started candidates.

+++ src/ast_grep_mcp/features/deduplication/config.py
@@ -63,9 +63,9 @@ class AnalysisConfig:
         exclude_patterns: Path patterns to exclude from analysis. Default: None
         parallel: Enable parallel execution for enrichment. Default: True
         max_workers: Maximum worker threads for parallel execution. Default: 4
-        timeout_per_candidate: Per-candidate enrichment timeout in seconds.
-            None uses the default (30s). Raise this if legitimate enrichment
-            operations are spuriously marked timed-out. Default: None
+        timeout_per_candidate: Per-candidate enrichment timeout in seconds
+            (float). None uses the default (30s). Raise this if legitimate
+            enrichment operations are spuriously marked timed-out. Default: None
         progress_callback: Optional callback for progress reporting.
             Signature: (stage_name: str, progress_percent: float) -> None
     """
@@ -86,7 +86,7 @@ class AnalysisConfig:
     # Parallel execution settings
     parallel: bool = True
     max_workers: int = ParallelProcessing.DEFAULT_WORKERS
-    timeout_per_candidate: Optional[int] = None
+    timeout_per_candidate: Optional[float] = None
 
     # Progress tracking
     progress_callback: Optional[Callable[[str, float], None]] = None
